fix: address pre-commit issues
