+++ tests/model_registry/image_validation/conftest.py
@@ -3,15 +3,12 @@ from typing import Any
 
 import pytest
 from kubernetes.dynamic import DynamicClient
-from ocp_resources.config_map import ConfigMap
 from ocp_resources.job import Job
 from ocp_resources.pod import Pod
 from pytest import FixtureRequest
-from pytest_testconfig import config as py_config
 
 from tests.model_registry.utils import get_latest_job_pod
 from utilities.general import wait_for_pods_by_labels
-from utilities.infra import ResourceNotFoundError
 
 
 @pytest.fixture(scope="class")
@@ -37,24 +34,6 @@ def resource_pods(request: FixtureRequest, admin_client: DynamicClient) -> list[
     label_selector = request.param.get("label_selector")
     assert namespace
     return list(Pod.get(namespace=namespace, label_selector=label_selector, client=admin_client))
-
-
-@pytest.fixture(scope="session")
-def async_upload_image(admin_client: DynamicClient) -> str:
-    """Async upload job image from the model-registry-operator-parameters ConfigMap."""
-    config_map = ConfigMap(
-        client=admin_client,
-        name="model-registry-operator-parameters",
-        namespace=py_config["applications_namespace"],
-    )
-
-    if not config_map.exists:
-        raise ResourceNotFoundError(
-            f"ConfigMap 'model-registry-operator-parameters' not found in"
-            f" namespace '{py_config['applications_namespace']}'"
-        )
-
-    return config_map.instance.data["IMAGES_JOBS_ASYNC_UPLOAD"]
 
 
 @pytest.fixture(scope="class")
