+++ tests/model_registry/model_registry/conftest.py
@@ -9,7 +9,6 @@ from kubernetes.dynamic import DynamicClient
 from kubernetes.dynamic.exceptions import ResourceNotFoundError
 from model_registry import ModelRegistry as ModelRegistryClient
 from model_registry.types import RegisteredModel
-from ocp_resources.config_map import ConfigMap
 from ocp_resources.deployment import Deployment
 from ocp_resources.namespace import Namespace
 from ocp_resources.pod import Pod
@@ -18,7 +17,6 @@ from ocp_resources.role_binding import RoleBinding
 from ocp_resources.service_account import ServiceAccount
 from pyhelper_utils.shell import run_command
 from pytest import FixtureRequest
-from pytest_testconfig import config as py_config
 
 from tests.model_registry.constants import (
     MODEL_REGISTRY_POD_FILTER,
@@ -240,37 +238,3 @@ def mr_access_role_binding(
         LOGGER.info(f"RoleBinding {binding.name} created successfully.")
         yield binding
         LOGGER.info(f"RoleBinding {binding.name} deletion initiated by context manager.")
-
-
-@pytest.fixture(scope="class")
-def async_upload_image(admin_client: DynamicClient) -> str:
-    """
-    Get the async upload image dynamically from the model-registry-operator-parameters ConfigMap.
-
-    This fetches the image from the cluster at runtime instead of using a hardcoded value.
-
-    Args:
-        admin_client: Kubernetes client for resource access
-
-    Returns:
-        str: The async upload image URL from the ConfigMap
-
-    Raises:
-        KeyError: If the ConfigMap or the required key doesn't exist
-    """
-    config_map = ConfigMap(
-        client=admin_client,
-        name="model-registry-operator-parameters",
-        namespace=py_config["applications_namespace"],
-    )
-
-    if not config_map.exists:
-        raise ResourceNotFoundError(
-            f"ConfigMap 'model-registry-operator-parameters' not found in"
-            f" namespace '{py_config['applications_namespace']}'"
-        )
-
-    try:
-        return config_map.instance.data["IMAGES_JOBS_ASYNC_UPLOAD"]
-    except KeyError as e:
-        raise KeyError(f"Key 'IMAGES_JOBS_ASYNC_UPLOAD' not found in ConfigMap data: {e}") from e
